change phone to numeric

+++ app/Http/Controllers/API/Customer/CustomerController.php
@@ -44,7 +44,7 @@ class CustomerController extends Controller
             'contact_person' => ['required', 'string'],
             'address' => ['required', 'string'],
             'email' => ['required', 'email'],
-            'phone' => ['required', 'integer'],
+            'phone' => ['required', 'numeric'],
         ]);
 
         $input = $request->all();
@@ -72,7 +72,7 @@ class CustomerController extends Controller
             'contact_person' => ['required', 'string'],
             'address' => ['required', 'string'],
             'email' => ['required', 'email'],
-            'phone' => ['required', 'integer'],
+            'phone' => ['required', 'numeric'],
         ]);
 
         $input = $request->all();
